feat : Lv 6. 달러 이외 통화를 환전할 수 있도록 수정

+++ src/main/java/com/sparta/currency_user/dto/ExchangeRequestDto.java
@@ -3,12 +3,14 @@ package com.sparta.currency_user.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class ExchangeRequestDto {
     private Long userId;
     private Long currencyId;
-    private double amountInKrw;
+    private BigDecimal amountInKrw;
 
 
 }
